+++ src/test/java/testclasses/RoomStubTest.java
@@ -1,20 +1,21 @@
-package saving;
+package testclasses;
 
 
 import core.Game;
+import core.StatusManager;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
-import saving.mockclasses.GameStub;
-import saving.mockclasses.PlayerSpy;
-import saving.mockclasses.RoomSpy;
-import saving.mockclasses.StatusManagerStub;
+import testclasses.mockclasses.GameStub;
+import testclasses.mockclasses.PlayerSpy;
+import testclasses.mockclasses.RoomStub;
+import testclasses.mockclasses.StatusManagerStub;
 
-class RoomTest {
+class RoomStubTest {
 
     @Test
     void enter_shouldCallOnEnterAndStartHandleUncleared() {
-        RoomSpy mockRoom = new RoomSpy(null, "Test Room");
+        RoomStub mockRoom = new RoomStub(null, "Test Room"); // Temporary null Game
         mockRoom.simulateUnclearedState();
         PlayerSpy mockPlayer = new PlayerSpy(mockRoom);
         GameStub gameMock = new GameStub(mockPlayer);
@@ -33,7 +34,7 @@ class RoomTest {
 
     @Test
     void enter_shouldCallOnEnterAndStartChooseRoom() {
-        RoomSpy mockRoom = new RoomSpy(null, "Test Room");
+        RoomStub mockRoom = new RoomStub(null, "Test Room"); // Temporary null Game
         mockRoom.simulateClearedState();
         PlayerSpy mockPlayer = new PlayerSpy(mockRoom);
         GameStub gameMock = new GameStub(mockPlayer);
@@ -53,8 +54,8 @@ class RoomTest {
     @Test
     void consume_shouldTrackInput() {
         Game gameMock = new GameStub();
-        RoomSpy room = new RoomSpy(gameMock, "Test Room");
-        RoomSpy neighbor = new RoomSpy(gameMock, "Neighbor");
+        RoomStub room = new RoomStub(gameMock, "Test Room");
+        RoomStub neighbor = new RoomStub(gameMock, "Neighbor");
         room.addNeighbor("up", neighbor);
 
         room.consume("up");
@@ -65,15 +66,14 @@ class RoomTest {
 
     @Test
     void handleUncleared_whenEntered() {
-        GameStub gameStub = new GameStub();
-        RoomSpy roomSpy = new RoomSpy(gameStub, "Test Room");
-        StatusManagerStub managerStub = new StatusManagerStub();
-        managerStub.setStatus(false);
-        gameStub.setTestStatusManager(managerStub);
-        roomSpy.simulateUnclearedState();
+        GameStub gameMock = new GameStub();
+        RoomStub room = new RoomStub(gameMock, "Test Room");
+        StatusManager manager = new StatusManagerStub();
+        gameMock.setTestStatusManager(manager);
+        room.simulateUnclearedState();
 
-        roomSpy.enter();
+        room.enter();
 
-        assertTrue(roomSpy.handleUnclearedCalled);
+        assertTrue(room.handleUnclearedCalled);
     }
 }
