+++ src/test/java/testclasses/RoomTest.java
@@ -1,21 +1,20 @@
-package saving;
+package testclasses;
 
 
 import core.Game;
-import core.StatusManager;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
-import saving.mockclasses.GameStub;
-import saving.mockclasses.PlayerSpy;
-import saving.mockclasses.RoomStub;
-import saving.mockclasses.StatusManagerStub;
+import testclasses.mockclasses.GameStub;
+import testclasses.mockclasses.PlayerSpy;
+import testclasses.mockclasses.RoomSpy;
+import testclasses.mockclasses.StatusManagerStub;
 
-class RoomStubTest {
+class RoomTest {
 
     @Test
     void enter_shouldCallOnEnterAndStartHandleUncleared() {
-        RoomStub mockRoom = new RoomStub(null, "Test Room"); // Temporary null Game
+        RoomSpy mockRoom = new RoomSpy(null, "Test Room");
         mockRoom.simulateUnclearedState();
         PlayerSpy mockPlayer = new PlayerSpy(mockRoom);
         GameStub gameMock = new GameStub(mockPlayer);
@@ -34,7 +33,7 @@ class RoomStubTest {
 
     @Test
     void enter_shouldCallOnEnterAndStartChooseRoom() {
-        RoomStub mockRoom = new RoomStub(null, "Test Room"); // Temporary null Game
+        RoomSpy mockRoom = new RoomSpy(null, "Test Room");
         mockRoom.simulateClearedState();
         PlayerSpy mockPlayer = new PlayerSpy(mockRoom);
         GameStub gameMock = new GameStub(mockPlayer);
@@ -54,8 +53,8 @@ class RoomStubTest {
     @Test
     void consume_shouldTrackInput() {
         Game gameMock = new GameStub();
-        RoomStub room = new RoomStub(gameMock, "Test Room");
-        RoomStub neighbor = new RoomStub(gameMock, "Neighbor");
+        RoomSpy room = new RoomSpy(gameMock, "Test Room");
+        RoomSpy neighbor = new RoomSpy(gameMock, "Neighbor");
         room.addNeighbor("up", neighbor);
 
         room.consume("up");
@@ -66,14 +65,15 @@ class RoomStubTest {
 
     @Test
     void handleUncleared_whenEntered() {
-        GameStub gameMock = new GameStub();
-        RoomStub room = new RoomStub(gameMock, "Test Room");
-        StatusManager manager = new StatusManagerStub();
-        gameMock.setTestStatusManager(manager);
-        room.simulateUnclearedState();
+        GameStub gameStub = new GameStub();
+        RoomSpy roomSpy = new RoomSpy(gameStub, "Test Room");
+        StatusManagerStub managerStub = new StatusManagerStub();
+        managerStub.setStatus(false);
+        gameStub.setTestStatusManager(managerStub);
+        roomSpy.simulateUnclearedState();
 
-        room.enter();
+        roomSpy.enter();
 
-        assertTrue(room.handleUnclearedCalled);
+        assertTrue(roomSpy.handleUnclearedCalled);
     }
 }
