+++ src/test/java/testclasses/mockclasses/GameStub.java
@@ -1,4 +1,4 @@
-package saving.mockclasses;
+package testclasses.mockclasses;
 
 import core.Game;
 import core.GameStatus;
