+++ src/test/java/testclasses/mockclasses/ScoreboardMock.java
@@ -1,4 +1,4 @@
-package saving.mockclasses;
+package testclasses.mockclasses;
 
 import core.Player;
 import observer.Scoreboard;
